When rdma_destroy_qp() and rdma_get_cm_event() are invoked in different
threads concurrently, there is a possibility that ibv_modify_qp() is
invoked on an already-destroyed qp inside the rdma_get_cm_event(), which
violates the thread-safety guarantee of rdmacm.

One possible workaround for user was to protect these two functions with
one mutex, but this would introduce an event_channel-level lock that is
shared among all the cm_id associated with the channel, which will lead
to high contention on this lock. So it is better to resolve this problem
inside rdmacm.

We take advantage of the cma_id_private.mut and use it to sync access to
qp inside these two functions.

Signed-off-by: Huaping Zhou <zhouhuaping.san@bytedance.com>