@FujiZ 2022-06-08
> AFAIK you can't multithread rdma_get_cm_event() with other CM operations, there is zero locking in this function so almost everything it does races against another cm operation of some kind.

I do some search in the mailing list of openfabrics and find this message which claims that librdmacm is thread safe, and the only exception is `rdma_migrate_id()`: https://lists.openfabrics.org/pipermail/general/2008-August/053908.html

> Apps certainly can't perform rdma_destroy_id() concurrently with get_event, I don't see why destroy_qp should be any different??

As it is noted in `man 3 rdma_get_cm_event`, "Destruction of an rdma_cm_id will block until related events have been acknowledged". If the concurrency between `rdma_destroy_id()` and `rdma_get_cm_event()` were not allowed, there would be no point in providing this guarantee since it would cause nothing but a deadlock if all cm operations must be serialized.