@rleon 2022-06-08
> > Apps certainly can't perform rdma_destroy_id() concurrently with get_event, I don't see why destroy_qp should be any different??
> 
> As it is noted in `man 3 rdma_get_cm_event`, "Destruction of an rdma_cm_id will block until related events have been acknowledged". If the concurrency between `rdma_destroy_id()` and `rdma_get_cm_event()` were not allowed, there would be no point in providing this guarantee since it would cause nothing but a deadlock if all cm operations must be serialized.

If I read the kernel code correctly, the rdma_destroy_id() separates between reported and not reported events.
https://elixir.bootlin.com/linux/latest/source/drivers/infiniband/core/ucma.c#L617
https://elixir.bootlin.com/linux/latest/source/drivers/infiniband/core/ucma.c#L566
https://elixir.bootlin.com/linux/latest/source/drivers/infiniband/core/ucma.c#L522

While reported event is not more than exit from ucma_get_event(), which is kernel part of rdma_get_cm_event().
https://elixir.bootlin.com/linux/latest/source/drivers/infiniband/core/ucma.c#L409

There is no protection from calling to rdma_destroy_id()->ucma_destroy_kern_id() in the middle of rdma_get_cm_event().