@FujiZ 2022-06-08
The resources associated with one cm_id consist of two parts: 1) user space part which is represented by `struct cma_id_private`; 2) kernel space part which is `struct ucma_context`. So destroying a cm_id involves destructing both parts properly.

What `ucma_destroy_id()` does is to destroy the kernel space part of cm_id, which involves cleaning up all unreported events associated with it. This is done by `ucma_cleanup_ctx_events()` (https://elixir.bootlin.com/linux/latest/source/drivers/infiniband/core/ucma.c#L522). Note that access to the `event_list` and `events_reported` is protected by the same mutex as `ucma_get_event()`.

After `ucma_destroy_id()` returns, it guarantees that the kernel part of the cm_id is destroyed, and no more events associated with it will be reported to user space through `ucma_get_event()`. However, the user space part can not be destroyed immediately, since there may still be some unacknowledged events in user space. So after that `rdma_destroy_id()` uses condition variable + mutex to wait until `events_completed == event_reported`, which means that all the events associated with this cm_id that are reported to user space are acknowledged by `rdma_ack_cm_event()`.

If `rdma_get_cm_event()` runs concurrently with `rdma_destroy_id()`, it is possible for `ucma_get_event()` to return an event associated with a cm_id whose kernel space part has already been freed. However, since the processing of event inside `rdma_get_cm_event()` only touches the user space part of the cm_id (`struct cma_id_private`), it doesn't matter if the kernel space part of it is destroyed or not. On the other hand, the user space part of one cm_id is guaranteed to be valid after the associating event is returned to user space through `ucma_get_event()`: since the `event_reported` counter in `struct ucma_context` is protected by a mutex, it is the exact number of events that are passed to user space. Thus, `rdma_destroy_id()` will block even if the associating unacknowledged event is just returned from `ucma_get_event()` and has not been returned to the caller of `rdma_get_cm_event()`.

To summarize, I think `rdma_get_cm_event()` and `rdma_destroy_id()` are thread safe.