@FujiZ 2022-06-10
Our application uses a background thread to handle connection setup procedure. It repeatedly calls `rdma_get_cm_event()`, initializes the necessary resource (QP, receive buffers, etc.) and handles state transitions of connections. Once the connection is established (by getting `ESTABLISHED` event), it will emit a message through a message queue to notify the target IO thread which the connection is bound to that this connection is ready to use. Since almost all the resource initialization and state transitions are handled in the background thread, they will not interfere with the ongoing IO operations in the IO thread, thus making the IO throughput & latency stable even when there are many connections to establish.

When the user wants to close a connection, it calls `rdma_destroy_qp()` and `rdma_destroy_id()` directly in the IO thread. Moreover, the user may want to close a connection before it is actually established (e.g. The connection setup procedure takes too long and a user-specified timer is fired). The resource of one connection can be easily protected by a mutex outside rdmacm so that even if they may be accessed by multiple threads, the operations applied to one cm_id can be serialized (e.g. `rdma_accept()` will not race with `rdma_destroy_qp()`). However, since `rdma_get_cm_event()` is outside the critical section formed by any particular mutex of the connection bound to the event channel, it may cause data race if QP is accessed inside `rdma_get_cm_event()` without any synchronization.

We have tried implementing two variants to work around this issue:

1. When the user wants to close a connection, it will emit a message from IO thread to background thread so that `rdma_destroy_qp()` and `rdma_destroy_id()` is scheduled to the background thread calling `rdma_get_cm_event()`.
2. Create one mutex for each event_channel and use it to serialize `rdma_get_cm_event()` with other rdmacm operations applied to those cm_ids associating with this channel.

We observed significant performance penalty in terms of the throughput and latency of connection setup in both variants: the average latency increases from 1.6ms to 13.7ms and 8.0ms respectively with one IO thread; and the latency increases from 8.3ms to 18.9ms and 19.8ms when there are 8 IO threads. And the throughput (RPS) of connection setup drops from 738 to 403 and 601, respectively. The performance of connection setup is crucial to the availability of our service.